adding remainder of workflows and update to style of site
