+++ docs/_data/navigation.yml
@@ -1,4 +1,6 @@
 - title: About
   url: about
+- title: Propoals
+  url: proposals
 - title: Documentation
   url: docs
