+++ docs/_layouts/page.html
@@ -3,5 +3,5 @@ layout: default
 ---
 {{ content }}
 {% include toc.html %}
-{% include permalinks.html %}
+{% unless page.skip_permalink %}{% include permalinks.html %}{% endunless %}
 {% include tags.html %}
