+++ docs/pages/index.md
@@ -29,5 +29,5 @@ to provide you or your organization an automated process to create and review pr
 
 {% endif %}
 
-For getting started with development, see the {% include doc.html name="Getting Started" path="getting-started" %} page. Would you like to request a feature or contribute? [Open an issue]({{ site.repo }}/issues)
+For getting started with working on proposals or development, see the {% include doc.html name="Getting Started" path="getting-started" %} page. Would you like to request a feature or contribute? [Open an issue]({{ site.repo }}/issues)
